requirements.txt
google-adk>=0.3.0
mcp>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
